moved default response handler to the root module because do not need anymore specialization for boa
