--- src/response_handler/boa.rs
@@ -1,8 +1,0 @@
-use crate::response_handler::{DefaultOutputter, DefaultResponse, ResponseHandler};
-
-pub struct DefaultResponseHandler;
-
-impl ResponseHandler for DefaultResponseHandler {
-    type Outputter = DefaultOutputter;
-    type Response = DefaultResponse;
-}
